Remove new warnings
ACCUMULO-4612 Remove new warning introduced when
Property.GENERAL_CLASSPATHS was deprecated.

ACCUMULO-3521 Remove warning about vestigial warnings suppression after
previous code cleanup

ACCUMULO-4501 Remove redundant cast to int and close() method call
leftover from summarization work

+++ core/src/main/java/org/apache/accumulo/core/summary/Gatherer.java
@@ -490,7 +490,7 @@ public class Gatherer {
   private int countFiles() throws TableNotFoundException, AccumuloException, AccumuloSecurityException {
     // TODO use a batch scanner + iterator to parallelize counting files
     Iterable<TabletMetadata> tmi = MetadataScanner.builder().from(ctx).overUserTableId(tableId, startRow, endRow).fetchFiles().fetchPrev().build();
-    return (int) StreamSupport.stream(tmi.spliterator(), false).mapToInt(tm -> tm.getFiles().size()).sum();
+    return StreamSupport.stream(tmi.spliterator(), false).mapToInt(tm -> tm.getFiles().size()).sum();
   }
 
   private class GatherRequest implements Supplier<SummaryCollection> {
